[Fix #1] Ensure forks don't attempt deploy in Travis CI
In this commit we're consolidating all of the logic that determines
when to run the deploy job into `.travis.yml` for clarity. We now
permit the deploy stage to run only if:

  - repo is "gonewest818/lein-sysutils"
  - job type is not a pull request
  - branch is master, or this is a tagged release

The purpose of this logic is to avoid attempting deploys to our
project Clojars account except when it is valid to do so.