modernize build matrix
Considering the state of clojure:
https://www.surveymonkey.com/results/SM-S9JVNXNQV/

...only 3.31% of the respondents use Clojure 1.7 or earlier,
suggesting we can reasonably drop that target from the build matrix.

However, 69.16% of respondents are still using Java 8, and due to
Oracle's end-of-life policies we can no longer install Oracle JDK 8 on
Travis build instances. We're switching to OpenJDK 8 in the build
matrix, and using what Oracle version we can use, which is v11.

Therefore the new build matrix is

Clojure: 1.8, 1.9, 1.10, and 1.11.0-master-SNAPSHOT
Java: openjdk{8,9,10,11} and oraclejdk11